automated-notes-generator/requirements.txt
streamlit
openai
python-dotenv
transformers
scikit-learn
pytube
moviepy
